PackageVersionEx refactor
